% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{create_tensor_cube}
\alias{create_tensor_cube}
\title{Create Tensor Cubes}
\usage{
create_tensor_cube(
  path_to_tiffs = "vignettes/data/switzerland",
  window_size = 30
)
}
\arguments{
\item{window_size}{}
}
\value{
torch tensor
}
\description{
Create Tensor Cubes
}
